functional: add mntent test
This only checks reading an fstab from an stream.  I haven't written
tests for setmntent(), addmntent(), or hasmntnent().

test_getmntent exposes a bug in musl 1.2.3 where lines omitting the
final two fields, which are supposed to be optional according to
fstab(5), are not accepted.  It also exposes a current bug, where if
the numeric fields are omitted, the last field seen will have a
newline appended to its value.  The tests all pass on Glibc.